Fix airflow worker grace period on Buster-based images
The problem was that PID1 (the tini process) was run as what ever the
"default" user was, which when run on our platform is currently
hard-coded to UID=100. We'd put in steps to switch to the other user to
run airflow (via sudo) but unfortunately this meant when we tried to
"gracefully shut down the celery worker" it would send the signal to
tini, which would try to forward it on to sudo, but this would fail
with:

    [FATAL tini (1)] Unexpected error when forwarding signal: 'Operation not permitted'

So the fix foe this is to make sure what ever is PID1 is running as the
right user.

The way we achieve this is:

1. Switch to `gosu` (it was already installed) as this doesn't leave a
   "sudo" process in the middle. If we kept sudo we would have sudo as
   PID1 which would handle signals properly (PID1 is special)
2. gosu doesn't have a config file -- you need to be root to be able to
   run it. To work around this we make it a sticky/SUID executable, so
   that anyone who can run it (via normal file permissions) will run it
   "as root"
3. To "limit the blast radius" of gosu, we only let root and members of
   the new "gosuers" group be able to execute the binary.

The end result of this: `tini` is PID1, and running as the "correct"
astro user, even the container was launched as UID 100, and signals
correctly propagate where needed