Fix entrypoints to correctly wait for DB/broker when no password (#59)
If you specify the celery broker URL as `redis://host:6379/0` it would
fail to correctly extract the host/port and would end up in a loop never actually run.

The problem was the awk pipeline required a username/password. So
instead replace it with a regex. Hopefully now we don't have two problems