Sync fab perms on web startup using airfl config
Because of the custom fab security manager
permissions need to be synced in Airflow on
webserver startup. This has up until now been
done in the `entrypoint.sh` script.

Now that we are actively promoting users
and enterprise environments to use our
core image, it will be expected for users
to quickly spin up our image in order to
test it out.

A workflow we can anticipate is for someone
to run the container interactively, create
a database and run the webserver. Since the
webserver will not have synced, the user
will be caught in an infite loop and will
likely become frustrated and abandon the
image, nor realizing that the `entrypoint.sh`
was clobered.

Easy option is to leverage the
`update_fab_perms` config instead.