Drop distro name from AC 2.2.0
closes https://github.com/astronomer/issues-airflow/issues/51

From AC 2.2.0, we have started using bullseye instead of buster. To be future proof we don't want to have distro name in image tag.

We can drop the if/else from jinja completely once we remove support for 1.10.10 and 1.10.12 as they are the last ones with alpine support.

Update config.yml.j2