Always build, test, and scan, and only require approval to push for version releases (#418)
**What this PR does / why we need it**:
closes https://github.com/astronomer/issues-airflow/issues/167
Alternative to #415.

This PR refactors the generated CircleCI config to always build, scan, and test images before pausing for approval to push images. For steps that build a versioned release of AC it preserves the CircleCI approvals and corresponding Slack approval notifications for pushing new images and the new build Slack notification to the QA channel. For dev and edge/`main` builds it skips the CircleCI/Slack approval process, but also keeps the new build Slack notifications to the QA channel.

**Special notes for your reviewer**:

The new build notifications to the QA channel may be too noisy. I'm happy to change this if necessary.